refactor: move run logic and context utilities to runlogic package
**Added:**

- Introduced `runlogic/run.go` with extracted run command logic, including `ExecuteRun`, context utilities (`WithConfig`, `WithViper`, etc.), and all supporting helpers for agent bundle, model invocation, diff application, and response handling

**Changed:**

- Updated `cmd/squad/root.go`, `cmd/squad/config.go`, and `cmd/squad/run.go` to use context utilities and run logic from the new `runlogic` package
- Refactored root command initialization into a `NewRootCmd` function in `root.go` for improved testability and clarity
- Replaced direct context key handling and run logic in `cmd/squad/run.go` with calls to `runlogic`
- Simplified Viper and config resolution to consistently use the refactored context helpers

**Removed:**

- Removed all run logic, agent bundle creation, model dispatch, diff application, and related context key handling from `cmd/squad/run.go` and `cmd/squad/root.go` now managed by `runlogic`
- Eliminated now-redundant context key types and helper functions from `cmd/squad/root.go` and `cmd/squad/run.go`
