+++ cmd/squad/run.go
@@ -24,65 +24,20 @@ package main
 
 import (
 	"bytes"
-	"context"
 	"fmt"
 	"io"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
-	"time"
 
-	"github.com/cowdogmoo/squad/agent"
-	"github.com/cowdogmoo/squad/logging"
-	"github.com/cowdogmoo/squad/ollama"
-	"github.com/cowdogmoo/squad/openairesponses"
-	"github.com/cowdogmoo/squad/tools"
+	"github.com/cowdogmoo/squad/runlogic"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	"github.com/tmc/langchaingo/llms"
-	"github.com/tmc/langchaingo/llms/anthropic"
-	"github.com/tmc/langchaingo/llms/openai"
 )
 
-// RunOptions holds the resolved configuration for a single run invocation.
-type RunOptions struct {
-	Agent             string
-	AgentsDir         string
-	WorkingDir        string
-	APIKey            string
-	BaseURL           string
-	Org               string
-	APIVersion        string
-	APIType           string
-	OpenAICompatMax   bool
-	Provider          string
-	Model             string
-	Temperature       float64
-	MaxTokens         int
-	System            string
-	Output            string
-	Print             bool
-	BundleOut         string
-	PrintBundle       bool
-	DryRun            bool
-	RequireActionable bool
-	Apply             bool
-	ApplyFallback     bool
-	NumCtx            int
-	Mode              string
-}
-
 // newRunOptions creates a RunOptions by reading resolved values from flags and Viper.
-func viperFromContext(cmd *cobra.Command) *viper.Viper {
-	if v, ok := cmd.Context().Value(viperKey).(*viper.Viper); ok {
-		return v
-	}
-	return viper.New()
-}
-
-func newRunOptions(cmd *cobra.Command) *RunOptions {
-	v := viperFromContext(cmd)
+func newRunOptions(cmd *cobra.Command) *runlogic.RunOptions {
+	v := runlogic.ViperFromContext(cmd)
 	agent := v.GetString("run.agent")
 	agentsDir := v.GetString("run.agents_dir")
 	workingDir := v.GetString("run.working_dir")
@@ -97,7 +52,7 @@ func newRunOptions(cmd *cobra.Command) *RunOptions {
 	applyFallback := v.GetBool("run.apply_fallback")
 	mode := v.GetString("run.mode")
 
-	return &RunOptions{
+	return &runlogic.RunOptions{
 		Agent:             agent,
 		AgentsDir:         agentsDir,
 		WorkingDir:        workingDir,
@@ -198,14 +153,11 @@ func newRunCmd() *cobra.Command {
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
 			opts := newRunOptions(cmd)
-			return executeRun(cmd, args, opts)
+			return runlogic.ExecuteRun(cmd, args, opts)
 		},
 	}
 
 	cmd.Flags().String("agent", "", "Agent name (e.g. go-cobra)")
-	if err := cmd.MarkFlagRequired("agent"); err != nil {
-		panic(err)
-	}
 	cmd.Flags().String("agents-dir", "", "Agents directory (default: ./agents, then ~/.config/squad/agents)")
 	cmd.Flags().String("working-dir", "", "Working directory (default: current working directory)")
 	cmd.Flags().String("api-key", "", "API key (overrides env/config)")
@@ -245,231 +197,6 @@ func newRunCmd() *cobra.Command {
 	return cmd
 }
 
-// executeRun contains the full run command logic, parameterized by RunOptions.
-func executeRun(cmd *cobra.Command, args []string, opts *RunOptions) error {
-	prompt, err := readPrompt(cmd, args)
-	if err != nil {
-		return err
-	}
-
-	workingDir, err := resolveWorkingDir(opts.WorkingDir)
-	if err != nil {
-		return err
-	}
-
-	bundle, err := prepareBundle(cmd, opts, prompt, workingDir)
-	if err != nil {
-		return err
-	}
-	if bundle == nil {
-		return nil // dry-run
-	}
-
-	ctx := tools.InitEdits(cmd.Context())
-	cmd.SetContext(ctx)
-	tools.ResetEditsApplied(ctx)
-	response, err := invokeModel(cmd, opts, bundle)
-	if err != nil {
-		return err
-	}
-
-	return handleResponse(cmd, opts, response, workingDir)
-}
-
-// prepareBundle builds the agent bundle and handles bundle output. Returns nil bundle for dry-run.
-func prepareBundle(cmd *cobra.Command, opts *RunOptions, prompt, workingDir string) (*agent.Bundle, error) {
-	agentsDir, err := resolveAgentsDir(opts.AgentsDir)
-	if err != nil {
-		return nil, err
-	}
-
-	bundle, err := agent.BuildBundle(agentsDir, opts.Agent, prompt, workingDir, opts.Mode)
-	if err != nil {
-		return nil, err
-	}
-
-	logging.InfoContext(cmd.Context(), "agent bundle ready (agent=%s provider=%s model=%s)", opts.Agent, opts.Provider, opts.Model)
-
-	if opts.PrintBundle {
-		if _, err := io.Copy(cmd.OutOrStdout(), bytes.NewReader(bundle.Combined)); err != nil {
-			return nil, err
-		}
-	}
-
-	if opts.BundleOut != "" {
-		if err := os.WriteFile(opts.BundleOut, bundle.Combined, 0o644); err != nil {
-			return nil, fmt.Errorf("failed to write bundle: %w", err)
-		}
-		logging.InfoContext(cmd.Context(), "bundle written to %s", opts.BundleOut)
-	}
-
-	if opts.DryRun {
-		return nil, nil
-	}
-
-	if configFromContext(cmd) == nil {
-		return nil, fmt.Errorf("config not available in context")
-	}
-
-	return bundle, nil
-}
-
-// invokeModel resolves provider settings and calls the appropriate model backend.
-func invokeModel(cmd *cobra.Command, opts *RunOptions, bundle *agent.Bundle) (string, error) {
-	v := viperFromContext(cmd)
-	provider := normalizeProvider(opts.Provider)
-	model := opts.Model
-	temperature := opts.Temperature
-	maxTokens := opts.MaxTokens
-
-	systemPrompt := bundle.System
-	if opts.System != "" {
-		systemPrompt += "\n\n## System Override\n\n" + strings.TrimSpace(opts.System) + "\n"
-	}
-
-	return callModel(cmd.Context(), v, provider, model, systemPrompt, bundle, temperature, maxTokens)
-}
-
-// handleResponse validates, applies, and writes the model response.
-func handleResponse(cmd *cobra.Command, opts *RunOptions, response, workingDir string) error {
-	if opts.RequireActionable {
-		if err := validateActionableResponse(cmd.Context(), response); err != nil {
-			return err
-		}
-	}
-
-	if opts.Apply {
-		if err := applyResponseDiff(cmd.Context(), response, workingDir, opts.ApplyFallback); err != nil {
-			return err
-		}
-	}
-
-	return writeResponse(cmd, response, opts)
-}
-
-// callModel dispatches the prompt to the appropriate model backend and returns the response.
-func callModel(ctx context.Context, v *viper.Viper, provider, model, systemPrompt string, bundle *agent.Bundle, temperature float64, maxTokens int) (string, error) {
-	if openairesponses.UseResponsesAPI(provider, model) {
-		return callResponsesAPI(ctx, v, model, systemPrompt, bundle, temperature, maxTokens)
-	}
-	return callLangChainLLM(ctx, v, provider, model, systemPrompt, bundle, temperature, maxTokens)
-}
-
-// callResponsesAPI runs the prompt via the OpenAI Responses API.
-func callResponsesAPI(ctx context.Context, v *viper.Viper, model, systemPrompt string, bundle *agent.Bundle, temperature float64, maxTokens int) (string, error) {
-	apiKey := v.GetString("provider.token")
-	if apiKey == "" {
-		apiKey = os.Getenv("OPENAI_API_KEY")
-	}
-	if apiKey == "" {
-		return "", fmt.Errorf("API key required for OpenAI Responses API: use --api-key, config provider.token, or OPENAI_API_KEY env var")
-	}
-
-	logging.InfoContext(ctx, "model call started via Responses API (model=%s)", model)
-	modelStart := time.Now()
-	response, err := openairesponses.RunWithTools(ctx, apiKey, v.GetString("provider.base_url"), model, systemPrompt, bundle.User, bundle.WorkDir, v.GetString("provider.organization"), temperature, maxTokens)
-	if err != nil {
-		return "", fmt.Errorf("model call failed: %w", err)
-	}
-	logging.InfoContext(ctx, "model call finished in %s (response-bytes=%d)", time.Since(modelStart).Round(time.Millisecond), len(response))
-	return response, nil
-}
-
-// callLangChainLLM runs the prompt via a LangChain-compatible LLM.
-func callLangChainLLM(ctx context.Context, v *viper.Viper, provider, model, systemPrompt string, bundle *agent.Bundle, temperature float64, maxTokens int) (string, error) {
-	llm, err := buildLLM(v, provider, model)
-	if err != nil {
-		return "", err
-	}
-
-	callOpts := buildCallOpts(v, provider, temperature, maxTokens)
-
-	logging.InfoContext(ctx, "model call started (provider=%s model=%s)", provider, model)
-	modelStart := time.Now()
-	response, err := tools.RunWithTools(ctx, llm, systemPrompt, bundle.User, bundle.WorkDir, callOpts...)
-	if err != nil {
-		return "", fmt.Errorf("model call failed: %w", err)
-	}
-	logging.InfoContext(ctx, "model call finished in %s (response-bytes=%d)", time.Since(modelStart).Round(time.Millisecond), len(response))
-	return response, nil
-}
-
-// buildCallOpts constructs LLM call options from provider settings.
-func buildCallOpts(v *viper.Viper, provider string, temperature float64, maxTokens int) []llms.CallOption {
-	callOpts := []llms.CallOption{
-		llms.WithTemperature(temperature),
-	}
-	if maxTokens <= 0 {
-		return callOpts
-	}
-	if !isOpenAICompatProvider(provider) {
-		return append(callOpts, llms.WithMaxTokens(maxTokens))
-	}
-	useLegacy := provider != "openai" || v.GetBool("provider.openai_compat_max_tokens")
-	if useLegacy {
-		return append(callOpts, llms.WithMaxTokens(maxTokens), openai.WithLegacyMaxTokensField())
-	}
-	return append(callOpts, openai.WithMaxCompletionTokens(maxTokens))
-}
-
-// applyResponseDiff extracts and applies a unified diff from the model response.
-func applyResponseDiff(ctx context.Context, response, workingDir string, fallback bool) error {
-	diff, err := extractUnifiedDiff(response)
-	if err != nil {
-		if responseIndicatesNoChanges(response) {
-			logging.InfoContext(ctx, "no changes reported; skipping apply")
-			return nil
-		}
-		if tools.EditsApplied(ctx) {
-			logging.InfoContext(ctx, "edits already applied via tools; skipping diff apply")
-			return nil
-		}
-		return err
-	}
-	if tools.EditsApplied(ctx) {
-		logging.InfoContext(ctx, "edits already applied via tools; skipping diff apply")
-		return nil
-	}
-	logging.InfoContext(ctx, "applying diff (%d bytes)", len(diff))
-	if err := applyUnifiedDiff(ctx, workingDir, diff, fallback); err != nil {
-		return err
-	}
-	logging.InfoContext(ctx, "diff applied to %s", workingDir)
-	return nil
-}
-
-// writeResponse outputs the model response to stdout and/or a file.
-func writeResponse(cmd *cobra.Command, response string, opts *RunOptions) error {
-	if opts.Print || opts.Output == "" {
-		if _, err := fmt.Fprintln(cmd.OutOrStdout(), response); err != nil {
-			return err
-		}
-	}
-	if opts.Output != "" {
-		if err := os.WriteFile(opts.Output, []byte(response), 0o644); err != nil {
-			return fmt.Errorf("failed to write response: %w", err)
-		}
-		logging.InfoContext(cmd.Context(), "response written to %s", opts.Output)
-	}
-	return nil
-}
-
-func readPrompt(cmd *cobra.Command, args []string) (string, error) {
-	if len(args) > 0 {
-		return strings.Join(args, " "), nil
-	}
-
-	input, err := io.ReadAll(cmd.InOrStdin())
-	if err != nil {
-		return "", fmt.Errorf("failed to read stdin: %w", err)
-	}
-	prompt := strings.TrimSpace(string(input))
-	if prompt == "" {
-		return "", fmt.Errorf("prompt is required (pass args or pipe stdin)")
-	}
-	return prompt, nil
-}
-
 func hasPipedInput(r io.Reader) bool {
 	if f, ok := r.(*os.File); ok {
 		fi, err := f.Stat()
@@ -482,213 +209,6 @@ func hasPipedInput(r io.Reader) bool {
 		return b.Len() > 0
 	}
 	return false
-}
-
-func resolveWorkingDir(dir string) (string, error) {
-	if dir == "" {
-		return os.Getwd()
-	}
-	return filepath.Abs(dir)
-}
-
-func resolveAgentsDir(explicit string) (string, error) {
-	if explicit != "" {
-		return filepath.Abs(explicit)
-	}
-
-	if stat, err := os.Stat("agents"); err == nil && stat.IsDir() {
-		return filepath.Abs("agents")
-	}
-
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return "", fmt.Errorf("failed to resolve agents dir: %w", err)
-	}
-	return filepath.Join(home, ".config", "squad", "agents"), nil
-}
-
-// buildLLM constructs an LLM model instance based on the provider and configuration.
-func buildLLM(v *viper.Viper, provider, model string) (llms.Model, error) {
-	provider = normalizeProvider(provider)
-	switch provider {
-	case "ollama":
-		return buildNativeOllamaLLM(v, model), nil
-	case "openai", "":
-		return buildOpenAICompatLLM(v, provider, model)
-	case "anthropic":
-		return buildAnthropicLLM(v, model)
-	default:
-		return nil, fmt.Errorf("provider not implemented: %s", provider)
-	}
-}
-
-func buildOpenAICompatLLM(v *viper.Viper, provider, model string) (llms.Model, error) {
-	oaiOpts := []openai.Option{}
-	if model != "" {
-		oaiOpts = append(oaiOpts, openai.WithModel(model))
-	}
-
-	if baseURL := v.GetString("provider.base_url"); baseURL != "" {
-		oaiOpts = append(oaiOpts, openai.WithBaseURL(baseURL))
-	}
-
-	if token := v.GetString("provider.token"); token != "" {
-		oaiOpts = append(oaiOpts, openai.WithToken(token))
-	}
-
-	if provider == "openai" || provider == "" {
-		if org := v.GetString("provider.organization"); org != "" {
-			oaiOpts = append(oaiOpts, openai.WithOrganization(org))
-		}
-
-		if apiVersion := v.GetString("provider.api_version"); apiVersion != "" {
-			oaiOpts = append(oaiOpts, openai.WithAPIVersion(apiVersion))
-		}
-
-		if apiType := strings.ToLower(v.GetString("provider.api_type")); apiType == "azure" {
-			oaiOpts = append(oaiOpts, openai.WithAPIType(openai.APITypeAzure))
-		}
-	}
-
-	return openai.New(oaiOpts...)
-}
-
-func buildAnthropicLLM(v *viper.Viper, model string) (llms.Model, error) {
-	aOpts := []anthropic.Option{}
-	if model != "" {
-		aOpts = append(aOpts, anthropic.WithModel(model))
-	}
-
-	if token := v.GetString("provider.token"); token != "" {
-		aOpts = append(aOpts, anthropic.WithToken(token))
-	}
-
-	if baseURL := v.GetString("provider.base_url"); baseURL != "" {
-		aOpts = append(aOpts, anthropic.WithBaseURL(baseURL))
-	}
-
-	return anthropic.New(aOpts...)
-}
-
-func normalizeProvider(provider string) string {
-	return strings.ToLower(strings.TrimSpace(provider))
-}
-
-func buildNativeOllamaLLM(v *viper.Viper, model string) llms.Model {
-	baseURL := v.GetString("provider.base_url")
-	if baseURL == "" {
-		baseURL = "http://localhost:11434"
-	}
-	numCtx := v.GetInt("provider.num_ctx")
-	if numCtx <= 0 {
-		numCtx = 32768
-	}
-	return ollama.New(baseURL, model, numCtx)
-}
-
-func isOpenAICompatProvider(provider string) bool {
-	return provider == "" || provider == "openai"
-}
-
-func validateActionableResponse(ctx context.Context, response string) error {
-	if tools.EditsApplied(ctx) {
-		return nil
-	}
-	if _, err := extractUnifiedDiff(response); err == nil {
-		return nil
-	}
-	lower := strings.ToLower(response)
-	if strings.Contains(lower, "files touched") || strings.Contains(lower, "no changes") {
-		return nil
-	}
-	return fmt.Errorf("response is not actionable: missing diff, files touched, or no changes section (disable with --require-actionable=false)")
-}
-
-func responseIndicatesNoChanges(response string) bool {
-	return strings.Contains(strings.ToLower(response), "no changes")
-}
-
-func extractUnifiedDiff(response string) (string, error) {
-	const fence = "```diff"
-	const altFence = "```patch"
-	const endFence = "```"
-	var blocks []string
-	for {
-		start := strings.Index(response, fence)
-		useFence := fence
-		if start == -1 {
-			start = strings.Index(response, altFence)
-			useFence = altFence
-		}
-		if start == -1 {
-			break
-		}
-		response = response[start+len(useFence):]
-		end := strings.Index(response, endFence)
-		if end == -1 {
-			block := strings.TrimSpace(response)
-			if block != "" {
-				blocks = append(blocks, block)
-			}
-			break
-		}
-		block := strings.TrimSpace(response[:end])
-		if block != "" {
-			blocks = append(blocks, block)
-		}
-		response = response[end+len(endFence):]
-	}
-	if len(blocks) == 0 {
-		return "", fmt.Errorf("apply requires a unified diff block (```diff ... ```)")
-	}
-	diff := strings.Join(blocks, "\n")
-	if !looksLikeDiff(diff) {
-		return "", fmt.Errorf("diff block does not look like a unified diff (missing diff headers)")
-	}
-	return diff, nil
-}
-
-func applyUnifiedDiff(ctx context.Context, workingDir, diff string, applyFallback bool) error {
-	if strings.TrimSpace(diff) == "" {
-		return fmt.Errorf("diff content is empty")
-	}
-
-	gitErr := applyWithGit(ctx, workingDir, diff)
-	if gitErr == nil {
-		return nil
-	}
-
-	if !applyFallback {
-		return fmt.Errorf("failed to apply diff with git: %v (hint: retry with --apply-fallback)", gitErr)
-	}
-
-	patchErr := applyWithPatch(ctx, workingDir, diff)
-	if patchErr == nil {
-		return nil
-	}
-
-	return fmt.Errorf("failed to apply diff with git or patch: %v; %v", gitErr, patchErr)
-}
-
-func looksLikeDiff(diff string) bool {
-	if strings.Contains(diff, "diff --git ") {
-		return true
-	}
-	if strings.Contains(diff, "--- a/") && strings.Contains(diff, "+++ b/") {
-		return true
-	}
-	return false
-}
-
-func applyWithGit(ctx context.Context, workingDir, diff string) error {
-	cmd := exec.CommandContext(ctx, "git", "apply", "--whitespace=nowarn", "--recount", "-")
-	cmd.Dir = workingDir
-	cmd.Stdin = strings.NewReader(diff)
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("git apply failed: %w: %s", err, strings.TrimSpace(string(out)))
-	}
-	return nil
 }
 
 func completeAgentNames(_ *cobra.Command, _ []string, toComplete string) ([]string, cobra.ShellCompDirective) {
@@ -711,15 +231,4 @@ func completeAgentNames(_ *cobra.Command, _ []string, toComplete string) ([]stri
 		}
 	}
 	return names, cobra.ShellCompDirectiveNoFileComp
-}
-
-func applyWithPatch(ctx context.Context, workingDir, diff string) error {
-	cmd := exec.CommandContext(ctx, "patch", "-p1")
-	cmd.Dir = workingDir
-	cmd.Stdin = strings.NewReader(diff)
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("patch failed: %w: %s", err, strings.TrimSpace(string(out)))
-	}
-	return nil
 }
